+++ app/src/docs/advanced/Internationalization.doc.tsx
@@ -1,11 +1,14 @@
 import * as React from 'react';
-import { EditableDocContent, BaseDocsBlock } from '../../common';
+import { EditableDocContent, BaseDocsBlock, DocExample } from '../../common';
 
 export class InternationalizationDoc extends BaseDocsBlock {
     title = 'Internationalization';
     renderContent() {
         return (
-            <EditableDocContent fileName="internationalization-intro" />
+            <>
+                <EditableDocContent fileName="internationalization-intro" />
+                <DocExample title="Right-to-left support(RTL)" path="./_examples/i18n/Rtl.example.tsx" />
+            </>
         );
     }
 }