+++ uui/components/datePickers/RangeDatePicker.tsx
@@ -7,12 +7,10 @@ import { Dropdown } from '@epam/uui-components';
 import { DropdownContainer } from '../overlays';
 import { FlexRow } from '../layout';
 import { RangeDatePickerBody } from './RangeDatePickerBody';
-import css from './RangeDatePicker.module.scss';
-import {
-    RangeDatePickerBodyValue, RangeDatePickerInputType, RangeDatePickerProps, RangeDatePickerValue,
-} from './types';
+import { RangeDatePickerBodyValue, RangeDatePickerInputType, RangeDatePickerProps, RangeDatePickerValue } from './types';
 import { defaultFormat, defaultRangeValue } from './helpers';
 import { RangeDatePickerInput } from './RangeDatePickerInput';
+import css from './RangeDatePicker.module.scss';
 
 const modifiers = [{
     name: 'offset',