+++ uui/components/pickers/index.ts
@@ -2,7 +2,7 @@ export * from './DataPickerBody';
 export * from './DataPickerFooter';
 export * from './DataPickerHeader';
 export * from './DataPickerRow';
-export * from './MobileDropdownWrapper';
+export * from './PickerBodyMobileView';
 export * from './PickerInput';
 export * from './PickerItem';
 export * from './PickerList';